% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova_with_Posttest.R,
%   R/Anova_with_Posttest_old.R
\name{Anova_with_Posttest}
\alias{Anova_with_Posttest}
\title{Anova_with_Posttest.}
\usage{
Anova_with_Posttest(data, value, group, method, level = 0.95)
}
\arguments{
\item{\code{method}}{Tukey or Duncan}
}
\value{
Return a datafram
}
\description{
\code{Anova_with_Posttest} Anova with Post Test.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
