% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calRTqPCR.R
\name{calRTqPCR}
\alias{calRTqPCR}
\title{Calculate Relative Expression value for RT-qPCR.}
\value{
Return a datafram
}
\description{
\code{calStandCurve} Calculate Standard Curve for RT-qPCR.
}
\examples{
filepath.1 <- system.file("examples", "20210928lx_1.txt", package = "pac4xiang")
filepath.2 <- system.file("examples", "20210929lx_1.txt", package = "pac4xiang")
stand.curve <- calStandCurve (data = filepath.1, genes = c("A","B","C","D"),
                              dilution = 4,start = 2, end = 7,rep = 4,
                              drop.NA = FALSE, fill.NA = "mean",
                              save.fig = TRUE, fig.type = 'pdf')
exp <- calRTqPCR(data = filepath.2,
                StandCurve = stand.curve,
                drop.NA = FALSE,
                fill.NA = "mean")
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
