% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calStandCurve.R
\name{calStandCurve}
\alias{calStandCurve}
\title{Calculate Standard Curve for RT-qPCR.}
\value{
Return a datafram
}
\description{
\code{calStandCurve} Calculate Standard Curve for RT-qPCR.
}
\examples{
filepath <- system.file("examples", "20210928lx_1.txt", package = "pac4xiang")
stand.curve <- calStandCurve (data = filepath, genes = c("A","B","C","D"),
                              dilution = 4,start = 2, end = 7,rep = 4,
                              drop.NA = FALSE, fill.NA = "mean",
                              save.fig = TRUE, fig.type = 'pdf')
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
