% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2fasta.R
\name{df2fasta}
\alias{df2fasta}
\title{Convert Data Frame Type File to Fasta.}
\value{
Return a datafram
}
\description{
\code{df2fasta} Convert Data Frame Type File to Fasta.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
