% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta2df.R
\name{fasta2df}
\alias{fasta2df}
\title{Convert Fasta Type File to Data Frame.}
\value{
Return a datafram
}
\description{
\code{fasta2df} Convert Fasta Type File to Data Frame.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
