% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignResults.R
\name{getAlignResults}
\alias{getAlignResults}
\title{Get Alignment Result from Clustalw.}
\value{
Return a datafram
}
\description{
\code{getAlignResults} Get Alignment Result from Clustalw.
}
\examples{
filepath <- system.file("examples", "pep.confirmed.aln", package = "pac4xiang")
aln.res <- calStandCurve (aln = filepath)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
