% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.aov.R
\name{mult.aov}
\alias{mult.aov}
\title{Multiple Group Anova.}
\value{
Return a datafram
}
\description{
\code{mult.aov} Multiple Group Anova.
}
\examples{
filepath <- system.file("examples", "df.1.txt", package = "pac4xiang")
aov.res <- mult.aov(data = filepath,
                    group1 = "group1",group2 = "group2",
                    value = "value", level = 0.95)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
