% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multGroupTtest.R
\name{multGroupTtest}
\alias{multGroupTtest}
\title{Multiple Group Student's-test.}
\value{
Return a datafram
}
\description{
\code{multGroupTtest} Multiple Group Student's-test.
}
\examples{
filepath <- system.file("examples", "df.1.txt", package = "pac4xiang")
t.res <- multGroupTtest(data = filepath,
                        group1 = "group1",group2 = "group2",
                        CK = "a",value = "value",level=0.95)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
