% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_corANDpvalue.R
\name{my_corANDpvalue}
\alias{my_corANDpvalue}
\title{group_df2list.}
\usage{
my_corANDpvalue(data, method = "pearson", cor = 0.6, pvalue = 0.05)
}
\description{
group_df2list.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
