% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot96well.R
\name{plot96well}
\alias{plot96well}
\title{Show 96-well Value of RT-qPCR.}
\value{
Return a plot
}
\description{
\code{plot96well} Show 96-well Value of RT-qPCR.
}
\examples{
filepath <- system.file("examples", "20210929lx_1.txt", package = "pac4xiang")
show96res <- plot96well(data = filepath)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
