% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCisElement.r
\name{plotCisElements}
\alias{plotCisElements}
\title{Plot Cis-element Location on Sequences.}
\value{
Return a plot
}
\description{
\code{plotCisElements} Plot Cis-element Location on Sequences.
}
\examples{
filepath <- system.file("examples", "plantCARE_output_PlantCARE_11359.tab", package = "pac4xiang")
plot.res <- plotCisElements(data = filepath,length = 1500)
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
