% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneStructure.R
\name{plotGeneStructure}
\alias{plotGeneStructure}
\title{Plot Gene Structure.}
\value{
Return a plot.
}
\description{
\code{plotGeneStructure} Plot Gene Structure.
}
\examples{
filepath <- system.file("examples", "gene.exon.info.gff", package = "pac4xiang")
gene.stru <- plotGeneStructure(gff = pathfile, tree = "none")
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
