% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme4Publication.R
\name{theme4Publication}
\alias{theme4Publication}
\title{ggplot theme.}
\usage{
theme4Publication(base_size = 14, base_family = "helvetica")
}
\description{
ggplot theme.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
