% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Anova_with_Posttest.R
\name{Anova_with_Posttest}
\alias{Anova_with_Posttest}
\title{Anova with Post Test.}
\arguments{
\item{\code{method}}{Tukey or Duncan}
}
\value{
Return a datafram
}
\description{
\code{Anova_with_Posttest} Anova with Post Test.
}
\author{
Xiang LI <lixiang117423@foxmail.com>
}
